% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.inc.R
\name{est.inc}
\alias{est.inc}
\title{Estimate incidence from cross-sectional data}
\usage{
est.inc(
  X,
  N0,
  mu_estimate,
  mu_CI_low,
  mu_CI_high,
  mu_CI_level = 0.95,
  prior = fit_gamma_distribution(mean = mu_estimate, lower = mu_CI_low, upper =
    mu_CI_high, confidence_level = mu_CI_level),
  mu_shape = prior[1],
  mu_scale = prior[2],
  incidence_CI_level = 0.95,
  alpha = 1 - incidence_CI_level,
  tol = 1e-08,
  max_incidence_CI_range = c(0, 100)
)
}
\arguments{
\item{X}{The number of MAA-positive participants in the cross-sectional sample}

\item{N0}{The number of seronegative participatns in the cross-sectional sample}

\item{mu_CI_low}{The lower confidence limit for the mean window period, mu (can skip this parameter if providing mu_shape and mu_scale)}

\item{mu_CI_high}{The upper confidence limit for the mean window period, mu (can skip this parameter if providing mu_shape and mu_scale)}

\item{mu_CI_level}{The confidence level for the mean window period, mu (can skip this parameter if providing mu_shape and mu_scale)}

\item{mu_shape}{The shape (gamma) parameter for the prior distribution of the mean window period, mu.}

\item{mu_scale}{The scale (beta) parameter for the prior distribution of the mean window period, mu.}

\item{incidence_CI_level}{The desired confidence interval coverage probability}

\item{alpha}{One minus the desired coverage probability}

\item{tol}{The desired tolerance for matching the specified confidence level}

\item{max_incidence_CI_range}{The range of incidence rate values to search over (if not wide enough, an error will be produced).}
}
\value{
a numeric vector containing the estimated incidence rate, lower confidence limit, and upper confidence limit, in units of new infections per 100 person-years.
}
\description{
Estimate incidence from cross-sectional data
}
