% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_sectional_incidence_estimates.R
\name{estimate_incidence}
\alias{estimate_incidence}
\title{Infer incidence rates from cross-sectional data}
\usage{
estimate_incidence(
  mu_table,
  mu_estimates = mu_table$mu,
  mu_CIs_lower = mu_table$"2.5\%",
  mu_CIs_upper = mu_table$"97.5\%",
  X,
  N,
  results = estimate_incidence_from_cohort(mu_estimate = mu_estimates, mu_CI_lower =
    mu_CIs_lower, mu_CI_upper = mu_CIs_upper, X = X, N = N, ...),
  results2 = round(results, 1),
  true_incidence = 1.9,
  digits = 1,
  ...
)
}
\arguments{
\item{mu_table}{}

\item{mu_estimates}{}

\item{mu_CIs_lower}{}

\item{mu_CIs_upper}{}

\item{X}{}

\item{N}{}

\item{results}{}

\item{results2}{}

\item{true_incidence}{}

\item{digits}{}

\item{...}{}
}
\value{

}
\description{
Infer incidence rates from cross-sectional data
}
