% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_squares.R
\name{sum_squares}
\alias{sum_squares}
\title{Calculate the error in matching a specified confidence interval with a Gamma distribution prior.}
\usage{
sum_squares(
  shape,
  lower_CI,
  upper_CI,
  mean,
  confidence_level = 0.95,
  alpha = (1 - confidence_level)/2
)
}
\arguments{
\item{shape}{The shape ("gamma") parameter of a Gamma distribution}

\item{lower_CI}{The lower confidence limit that we are trying to match}

\item{upper_CI}{The upper confidence limit that we are trying to match}

\item{mean}{The mean of the Gamma distribution (= shape * scale)}

\item{confidence_level}{The confidence level of the confidence interval (1 - alpha)}

\item{alpha}{= 1 - `confidence_level`}
}
\value{
The sum of relative squared errors between the confidence limits and the quantiles of the fitted distribution.
}
\description{
Calculate the error in matching a specified confidence interval with a Gamma distribution prior.
}
