% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_squares.R
\name{sum_squares}
\alias{sum_squares}
\title{Title}
\usage{
sum_squares(
  shape,
  lower_CI,
  upper_CI,
  mean,
  confidence_level = 0.95,
  alpha = (1 - confidence_level)/2
)
}
\arguments{
\item{shape}{}

\item{lower_CI}{}

\item{upper_CI}{}

\item{mean}{}

\item{confidence_level}{}

\item{percentile}{}
}
\value{

}
\description{
Title
}
